cricket partially completed

+++ CoreJava_Dec18/src/com/vtalent/hanu/classtask/Match.java
@@ -2,7 +2,7 @@ package com.vtalent.hanu.classtask;
 
 interface Match {
 
-	public double score(Team team);
+	public int score(Team team);
 	
 	public String winner(Team t1,Team t2);
 }
